voice_training/requirements.txt
httpx>=0.27.0
anthropic>=0.40.0
openai-whisper>=20231117
pydub>=0.25.1
